- firmware:     - added support for serial_gateway on comports 1, 2, 3     - increased serial buffer size from 100 to 270 bytes.
- completed the rough MinilabProtocol script (can successfully exercise all commands from the board)

+++ firmware/proto.h
@@ -13,9 +13,9 @@ extern "C" {
 // -1 when request receive-parse error
 //  0 if nothing considerable has changed
 //  1 when valid request received
-i8   proto_try_read_request ( Byte* buffer, u8 *len, u8 max_size );
-void proto_send_reply       (u8 *buffer, u8 length);
-void proto_send_error_reply (u8 code);
+i8   proto_try_read_request ( Byte* buffer, u16 *len, u16 max_size );
+void proto_send_reply       ( u8 *buffer, u16 length );
+void proto_send_error_reply ( u8 code );
 
 
 #ifdef __cplusplus
